ft_printf

+++ ft_printf/ft_putpointer.c
@@ -6,20 +6,21 @@
 /*   By: pvass <pvass@student.42.fr>                +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/04/13 20:25:59 by pvass             #+#    #+#             */
-/*   Updated: 2024/04/15 11:44:22 by pvass            ###   ########.fr       */
+/*   Updated: 2024/04/15 17:47:01 by pvass            ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
-#include "libft.h"
+#include "./includes/libft.h"
+#include "./includes/ft_printf.h"
 
-static void	ft_putp(unsigned long long ptr)
+static void	ft_putp(uintptr_t ptr)
 {
 	if (ptr >= 16)
 	{
 		ft_putp(ptr / 16);
 		ft_putp(ptr % 16);
 	}
-	else 
+	else
 	{
 		if (ptr <= 9)
 			ft_putchar_fd(ptr + '0', 1);
@@ -28,9 +29,9 @@ static void	ft_putp(unsigned long long ptr)
 	}
 }
 
-static int	ptr_len(unsigned long long ptr)
+static int	ptr_len(uintptr_t ptr)
 {
-	int i;
+	int		i;
 
 	i = 0;
 	while (ptr > 0)
@@ -41,7 +42,7 @@ static int	ptr_len(unsigned long long ptr)
 	return (i);
 }
 
-int	ft_putpointer (unsigned long long ptr)
+int	ft_putpointer(uintptr_t ptr)
 {
 	int	len;
